requirements-dev.txt
black
flake8
flask-debugtoolbar
flask-shell-ipython
ipdb
ipython
isort
pytest
pytest-flask
pytest-cov
honcho
pre-commit
factory-boy